# [cluster-test] Added perf measurement test