As a start point of "better performance measurement" https://github.com/diem/diem/issues/8023. This test can make perf test push work load stage by stage.

arguments introduced:
--base-tps: start workload for a fixed tps number
--step-length: incremental number of workload for each step
--step-num: step numbers for whole perf test

## Motivation

(Write your motivation for proposed changes here.)

### Have you read the [Contributing Guidelines on pull requests](https://github.com/diem/diem/blob/main/CONTRIBUTING.md#pull-requests)?

(Write your answer here.)

## Test Plan

./scripts/cti --pr 8292 --run measure -- --base-tps 500 --step-length 100 --step-num 3

```
Starting experiment Perf Measurement: start tps 500, steps number 3, step length = 100
Experiment Result: Test runner setup time spent 249 secs
Step 0 : 385 TPS, 1338 ms latency, 1500 ms p99 latency, no expired txns
Step 1 : 429 TPS, 1488 ms latency, 1600 ms p99 latency, no expired txns
Step 2 : 462 TPS, 1577 ms latency, 1800 ms p99 latency, no expired txns
```

## Related PRs

(If this PR adds or changes functionality, please take some time to update or suggest changes to the docs at https://developers.diem.com, and link to your PR here.)

## If targeting a release branch, please fill the below out as well

 * Justification and breaking nature (who does it affect? validators, full nodes, tooling, operators, AOS, etc.)
 * Comprehensive test results that demonstrate the fix working and not breaking existing workflows.
 * Why we must have it for V1 launch.
 * What workarounds and alternative we have if we do not push the PR.